+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.1.0
+- added support for root IDs
+- added Explorer context validation
+- added Explorer context build
+- added Editor tab XML format
+
 ## 0.0.2
 - code cleanup
 ## 0.0.1